+++ src/ReleaseService.php
@@ -3,6 +3,8 @@
 namespace TPG\Attache;
 
 use Illuminate\Support\Arr;
+use Illuminate\Support\Str;
+use TPG\Attache\Exceptions\ProcessException;
 use TPG\Attache\Exceptions\ServerException;
 
 class ReleaseService
@@ -49,9 +51,16 @@ class ReleaseService
 
     public function hasInstallation(): bool
     {
-        $releases = $this->getReleases();
+        try {
+            $releases = $this->getReleases();
+            return count($releases) > 0;
+        } catch (\Exception $e) {
+            if (! Str::contains(strtolower($e->getMessage()), 'no such file or directory')) {
+                throw $e;
+            }
 
-        return count($releases) > 0;
+            return false;
+        }
     }
 
     /**
@@ -85,7 +94,7 @@ class ReleaseService
         return Arr::get($matches, 'id');
     }
 
-    protected function run(Task $task): string
+    protected function run(Task $task): ?string
     {
         $data = null;